Start on some plumbing
Made some progress on the find_payment_method api call support.  This is
driving support for some of underlying https interactions.

The approach has much in common with Active Merchant.  In fact a good
bit of it is simply copied and potentially simplified for our use.

Much more work to do.