Improve error handling
Now the exceptions raised by errors on the Spreedly side remember the
entire list of errors that were returned.  It turns out that there are
times when multiple errors can be returned when a transaction can't be
created.

One example is the API call to create a credit card.  The errors in this
case could include all of the validation errors.

The message in this case is the concatenation of the error messages.
Not sure if this is the best way to do this.