Reorganize response stubs
Most of the time, I don't want to be looking at the stubbed out response
bodies when messing with unit tests.  And they have a tendency to take
up a huge percentage of a unit test's file.  So, I thought I'd
experiment by pulling them out into modules.

Not sure yet whether this will be annoying or wonderful.