Clean up assertion weirdness
The unit test for fields is already testing that boolean fields are
really returning a boolean and not a string and that the ? accessors are
also defined appropriately.  We don't need to test that everywhere.
Instead, just do a simple assertion like most folks will use it in
practice.

HT: @ntalbott