Fixes to a bunch of the remote tests.
Some of the tests are destined to fail, since they rely on the environment already having many gateways, transactions, etc in them. If you are using a test environment, particularly one that was just created to run these tests through, they will fail.

The lists tests are a bit of a mystery. I know the functionality works, i.e. filter by a token to return a list of items after that token, but it is not working in the tests.

Perhaps other people aren't running these